Added dependencies provision, replaced relative paths with absolute ones and some other minor fixes

+++ utils/get_res.sh
@@ -2,9 +2,26 @@
 # SPDX-FileCopyrightText: © 2023 Project's authors 
 # SPDX-License-Identifier: MIT
 
-utils_path="$(dirname ${0})"
-res_path="${utils_path}/../res"
-svn export https://github.com/PokeAPI/sprites/trunk/sprites/pokemon/versions/generation-iv/platinum "${res_path}"
+dependencies="svn curl jq mogrify"
+missing_dependencies=""
+for dependency in ${dependencies}; do
+    if ! command -v "${dependency}" &>/dev/null; then
+        if [ -n "${missing_dependencies}" ]; then
+            missing_dependencies="${missing_dependencies}, ${dependency}"
+        else
+            missing_dependencies="${dependency}"
+        fi
+    fi
+done
+
+if [ -n "${missing_dependencies}" ]; then
+    printf "\e[31;1mError: The following dependencies are not in PATH: ${missing_dependencies}\e[0m"
+    exit 1
+fi
+
+src_path="$(readlink -f "$(dirname "${0}")")"
+res_path="$(readlink -f "$(dirname "${src_path}")")/res"
+svn export --force https://github.com/PokeAPI/sprites/trunk/sprites/pokemon/versions/generation-iv/platinum "${res_path}"
 rm -rf "${res_path}/back" "${res_path}/shiny" "${res_path}/female"
 
 # Keep only one Unown form
@@ -33,8 +50,8 @@ rm "${res_path}"/492-*
 rm "${res_path}"/493-*
 
 for id in {1..493}; do
-    printf "${id}/493\n"
-    name="$(curl "https://pokeapi.co/api/v2/pokemon/${id}" | jq ".name")"
+    printf "\e[35;1m${id}/493\e[0m\n"
+    name="$(curl "https://pokeapi.co/api/v2/pokemon/${id}" | jq -r ".name")"
     mv "${res_path}/${id}.png" "${res_path}/${id}-${name^}.png"
 done 
 
@@ -45,8 +62,11 @@ mv "${res_path}/487-Giratina-altered.png" "${res_path}/487-Giratina.png"
 mv "${res_path}/492-Shaymin-land.png" "${res_path}/492-Shaymin.png"
 
 # Trim transparency
-mogrify -trim ${res_path}/*
+mogrify -trim "${res_path}/*"
 
+unset dependencies
+unset missing_dependencies
+unset dependency
 unset utils_path
 unset res_path
 unset name
